Castlebase fixes/adjustments + default.dae with Stalker hole WIP
- castlebase_wall.dds:
Added new row of bricks so that the mentioned white spots at the bottom are not visible when seen from above.

- castlebase_wall_height_hole.dds:
Slightly modified the height map so that the hole sides don't look like they are floating above brick gaps.

- castlebase_wall_hole.mat:
Made alternate material which includes this height map, for the .dae with the hole.

- default.dae:
A few fixes, mostly regarding seemingly messed up UV mapping (resulting in a few pitch black faces, and a clear seam at the border connecting with another default wall, in the bottom side border).

- default_ghoulhole.dae:
(WIP) missing hole collisions, not final texture. Castlebase default.dae featuring a hole for the stalker to go in.