Greenhouse, storage & both entrance
Greenhouse
Added broken lamps in the map, one not broken in the greenhouse area.
Added cables to all broken lamps

Storage
Changed all Illum light to match the the rest of the color of the lamps.

Entrance (both versions)
Made sure that the storage transfers corridors had the same Illum light as the once at storage map