Fixed crash when saving game at Entrance map
The cause was an array referencing pickables entities directly which was saved outside functions.

When these entities got picked up, they stopped existing so the reference probably was null and for some reason when saving the game it was probably saving something this array or something related and caused a crash.